fix(ci): restore shared worktree contracts
