+++ crates/agent-gui/src/lib/settings/index.ts
@@ -22,6 +22,7 @@ import {
   MAX_CHAT_TRANSCRIPT_WIDTH,
   MIN_CHAT_TRANSCRIPT_WIDTH,
 } from "@liveagent/ui/lib/transcript-width/transcriptWidthModel";
+import type { WorkspaceProjectGroup } from "@liveagent/ui/lib/workspaceProjectTypes";
 import { DEFAULT_LOCALE, type Locale, normalizeLocale } from "../../i18n/config";
 import {
   ANTHROPIC_LONG_CONTEXT_WINDOW,
@@ -34,6 +35,7 @@ import {
 } from "../providers/anthropicModels";
 import { normalizeFontFamily } from "../system/fontFamily";
 
+export type { WorkspaceProjectGroup } from "@liveagent/ui/lib/workspaceProjectTypes";
 export { normalizeFontFamily } from "../system/fontFamily";
 
 export function isThinkingAlwaysOnForModel(
@@ -296,23 +298,6 @@ export type EffectiveWorkspaceResources = {
 };
 
 export type WorkspaceProjectKind = "managed" | "folder" | "history";
-
-/**
- * 侧边栏项目分组。成员用原始路径存储（匹配时经
- * `workspaceProjectPathKey` 归一化），与 hidden/missing/archived 一致。
- *
- * `sourceProjectPath` 标记自动分组（git worktree 聚合）：指向原始仓库
- * 项目的路径，重命名分组后仍可据此复用，避免重复建组。
- */
-export type WorkspaceProjectGroup = {
-  id: string;
-  name: string;
-  projectPaths: string[];
-  sourceProjectPath?: string;
-  collapsed?: boolean;
-  createdAt: number;
-  updatedAt: number;
-};
 
 export type WorkspaceProject = {
   id: string;
