+++ crates/agent-ui/src/lib/git/types.ts
@@ -202,8 +202,8 @@ export type GitClient = {
   push(workdir: string): Promise<GitOperationResponse>;
   deleteBranch(workdir: string, branch: string, force?: boolean): Promise<GitOperationResponse>;
   renameBranch(workdir: string, branch: string, newBranch: string): Promise<GitOperationResponse>;
-  createWorktree(workdir: string, name: string, startPoint?: string): Promise<GitWorktreeResponse>;
-  removeWorktree(
+  createWorktree?(workdir: string, name: string, startPoint?: string): Promise<GitWorktreeResponse>;
+  removeWorktree?(
     workdir: string,
     worktreePath: string,
     force?: boolean,
